+++ krust/tests/ga_solver_tests.rs
@@ -11,7 +11,7 @@ mod solver_tests {
     use krust::matrices::{IDENTITY};
     use krust::solver_ga::IKSolverGA;
 
-    const TARGET: Matrix4<f64> = Matrix4::new(  
+    const TARGET: Matrix4<f32> = Matrix4::new(  
         1.0,0.0,0.0,0.0,
         0.0,1.0,0.0,3.0,
         0.0,0.0,1.0,3.0,
@@ -22,9 +22,9 @@ mod solver_tests {
     fn test_new_solver_success() {
 
         // Small case
-        let angles: Vec<f64> = vec![0.0,0.0,0.0];
-        let axes: Vec<Vector3<f64>> = vec![*Vector3::x_axis(), *Vector3::y_axis(), *Vector3::z_axis()];
-        let radii: Vec<f64> = vec![5.0,3.0,1.0];
+        let angles: Vec<f32> = vec![0.0,0.0,0.0];
+        let axes: Vec<Vector3<f32>> = vec![*Vector3::x_axis(), *Vector3::y_axis(), *Vector3::z_axis()];
+        let radii: Vec<f32> = vec![5.0,3.0,1.0];
 
         let ik_solver: IKSolverGA = IKSolverGA::new(IDENTITY, &angles, &axes, &radii);
 
@@ -40,9 +40,9 @@ mod solver_tests {
     fn test_generate_matrices_fail() {
 
         // Test with unequal input vectors
-        let angles: Vec<f64> = vec![0.0,0.0];
-        let axes: Vec<Vector3<f64>> = vec![*Vector3::x_axis(), *Vector3::y_axis(), *Vector3::z_axis()];
-        let radii: Vec<f64> = vec![5.0,3.0,1.0];
+        let angles: Vec<f32> = vec![0.0,0.0];
+        let axes: Vec<Vector3<f32>> = vec![*Vector3::x_axis(), *Vector3::y_axis(), *Vector3::z_axis()];
+        let radii: Vec<f32> = vec![5.0,3.0,1.0];
 
         IKSolverGA::new(IDENTITY, &angles, &axes, &radii);
 
@@ -51,9 +51,9 @@ mod solver_tests {
     #[test]
     fn test_solver_run() {
 
-        let angles: Vec<f64> = vec![0.0,0.0,0.0];
-        let axes: Vec<Vector3<f64>> = vec![*Vector3::x_axis(), *Vector3::x_axis(), *Vector3::x_axis()];
-        let radii: Vec<f64> = vec![2.0,2.0,2.0];
+        let angles: Vec<f32> = vec![0.0,0.0,0.0];
+        let axes: Vec<Vector3<f32>> = vec![*Vector3::x_axis(), *Vector3::x_axis(), *Vector3::x_axis()];
+        let radii: Vec<f32> = vec![2.0,2.0,2.0];
 
         let mut ik_solver: IKSolverGA = IKSolverGA::new(IDENTITY, &angles, &axes, &radii);
 
@@ -68,9 +68,9 @@ mod solver_tests {
     #[test]
     fn test_solver_solve() {
 
-        let angles: Vec<f64> = vec![0.0,0.0,0.0];
-        let axes: Vec<Vector3<f64>> = vec![*Vector3::x_axis(), *Vector3::x_axis(), *Vector3::x_axis()];
-        let radii: Vec<f64> = vec![2.0,2.0,2.0];
+        let angles: Vec<f32> = vec![0.0,0.0,0.0];
+        let axes: Vec<Vector3<f32>> = vec![*Vector3::x_axis(), *Vector3::x_axis(), *Vector3::x_axis()];
+        let radii: Vec<f32> = vec![2.0,2.0,2.0];
 
         let mut ik_solver: IKSolverGA = IKSolverGA::new(IDENTITY, &angles, &axes, &radii);
